+++ ribs/emitters/_iso_line_emitter.py
@@ -3,7 +3,6 @@ import numpy as np
 
 from ribs._utils import check_batch_shape, check_shape, np_scalar
 from ribs.emitters._emitter_base import EmitterBase
-from ribs.emitters.operators import IsoLineOperator
 
 
 class IsoLineEmitter(EmitterBase):
@@ -69,6 +68,7 @@ class IsoLineEmitter(EmitterBase):
             bounds=bounds,
         )
 
+        self._rng = np.random.default_rng(seed)
         self._batch_size = batch_size
         self._iso_sigma = np_scalar(iso_sigma, dtype=archive.dtypes["solution"])
         self._line_sigma = np_scalar(line_sigma, archive.dtypes["solution"])
@@ -90,10 +90,6 @@ class IsoLineEmitter(EmitterBase):
                 initial_solutions, dtype=archive.dtypes["solution"])
             check_batch_shape(self._initial_solutions, "initial_solutions",
                               archive.solution_dim, "archive.solution_dim")
-
-        self._operator = IsoLineOperator(iso_sigma=self._iso_sigma,
-                                         line_sigma=self._line_sigma,
-                                         seed=seed)
 
     @property
     def x0(self):
@@ -154,12 +150,28 @@ class IsoLineEmitter(EmitterBase):
             return self._clip(self.initial_solutions)
 
         if self.archive.empty:
+            # Note: Since the parents are all `x0`, the `directions` below will
+            # all be 0.
             parents = np.repeat(self.x0[None],
-                                repeats=2 * self._batch_size,
+                                repeats=2 * self.batch_size,
                                 axis=0)
         else:
             parents = self.archive.sample_elites(2 *
-                                                 self._batch_size)["solution"]
+                                                 self.batch_size)["solution"]
 
-        return self._clip(
-            self._operator.ask(parents.reshape(2, self._batch_size, -1)))
+        parents = parents.reshape(2, self.batch_size, self.solution_dim)
+        elites = parents[0]
+        directions = parents[1] - parents[0]
+
+        iso_gaussian = self._rng.normal(
+            scale=self.iso_sigma,
+            size=(self.batch_size, self.solution_dim),
+        ).astype(elites.dtype)
+        line_gaussian = self._rng.normal(
+            scale=self.line_sigma,
+            size=(self.batch_size, 1),
+        ).astype(elites.dtype)
+
+        solutions = elites + iso_gaussian + line_gaussian * directions
+
+        return self._clip(solutions)